Working on xcode, but not on swift playground(ipad)

+++ BinarySafe.playground/Sources/GeometryData.swift
@@ -2,5 +2,5 @@ import SpriteKit
 
 struct GeometryData {
     static let fullRadianOfCircle: CGFloat = 6.28318530
-    static let fullDegreeOfCircle: Int = 360
+    static let fullDegreeOfCircle: Double = 360
 }
